estimate_hcru

This function calculates estimates of healthcare resource utilization (HCRU) from electronic health record data across various care settings (e.g., IP, OP, ED/ER). It provides descriptive summaries of patient counts, encounters, costs, length of stay, and readmission rates for pre- and post-index periods.

Usage

estimate_hcru(
  data,
  cohort_col = "cohort",
  patient_id_col = "patient_id",
  admit_col = "admission_date",
  discharge_col = "discharge_date",
  index_col = "index_date",
  visit_col = "visit_date",
  encounter_id_col = "encounter_id",
  setting_col = "care_setting",
  cost_col = "cost_usd",
  readmission_col = "readmission",
  time_window_col = "period",
  los_col = "length_of_stay",
  custom_var_list = NULL,
  pre_days = 180,
  post_days = 365,
  readmission_days_rule = 30,
  group_var_main = "cohort",
  group_var_by = "care_setting",
  test = NULL,
  timeline = "Pre",
  gt_output = TRUE
)

Arguments

data

A dataframe specifying the health care details.

cohort_col

A character specifying the name of the cohort column.

patient_id_col

A character specifying the name of the patient identifier column.

admit_col

A character specifying the name of the date of admission column.

discharge_col

A character specifying the name of the date of discharge column.

index_col

A character specifying the name of the index date or diagnosis column.

visit_col

A character specifying the name of the date of visit/claim column.

encounter_id_col

A character specifying the name of the encounter/claim column.

setting_col

A character specifying the name of the HCRU setting column e.g. IP, ED, OP, etc.

cost_col

A character specifying the name of cost column.

readmission_col

A character specifying the name of readmission column.

time_window_col

A character specifying the name of time window column.

los_col

A character specifying the name of length of stay column.

custom_var_list

A character vector providing the list of additional columns.

pre_days

Number of days before index (default 180 days).

post_days

Number of days after index (default 365 days).

readmission_days_rule

Rule for how many days can be permissible to define readmission criteria in AP setting (default 30 days).

group_var_main

A character specifying the name of the main grouping column.

group_var_by

A character specifying the name of the secondary grouping column.

test

An optional named list of statistical tests (e.g., list(age = "wilcox.test")).

timeline

A character specifying the timeline window (default "Pre").

gt_output

Logical; if TRUE, also returns output formatted using gtsummary (default is TRUE).

Value

A list containing one or two summary data frames:

Summary by settings using dplyr

A descriptive summary of HCRU metrics by cohort, setting, and time window.

Summary by settings using gtsummary (optional)

Formatted summary statistics using gtsummary, if gt_output = TRUE.

Examples

# \donttest{
df <- hcru_sample_data[sample(nrow(hcru_sample_data), 10), ]
estimate_hcru(data = df)
#> Warning: The `update` argument of `modify_header()` is deprecated as of gtsummary 2.0.0.
#> ℹ Use `modify_header(...)` input instead. Dynamic dots allow for syntax like
#>   `modify_header(!!!list(...))`.
#> ℹ The deprecated feature was likely used in the gtsummary package.
#>   Please report the issue at <https://github.com/ddsjoberg/gtsummary/issues>.
#> $`Summary by settings using dplyr`
#> # A tibble: 9 × 15
#>   patient_id cohort    care_setting time_window  Days Month   Year Visits   Cost
#>   <chr>      <chr>     <chr>        <chr>       <dbl> <dbl>  <dbl>  <dbl>  <dbl>
#> 1 P_1015     control   OP           Pre            14 0.460 0.0383      1   587.
#> 2 P_1038     control   OP           Pre            82 2.70  0.224       1   175.
#> 3 P_1085     control   ED           Post           87 2.86  0.238       1  4572.
#> 4 P_1090     treatment ED           Pre             5 0.164 0.0137      1  4885.
#> 5 P_1096     treatment OP           Post            8 0.263 0.0219      1   775.
#> 6 P_1100     treatment OP           Post           80 2.63  0.219       1  1882.
#> 7 P_1105     treatment ED           Pre            32 1.05  0.0876      1  1312.
#> 8 P_1164     treatment IP           Post           85 2.79  0.233       1 16170.
#> 9 P_1170     treatment IP           Post          155 5.10  0.424       1  7040.
#> # ℹ 6 more variables: LOS <dbl>, Readmission <dbl>, Visit_PPPM <dbl>,
#> #   Visit_PPPY <dbl>, Cost_PPPM <dbl>, Cost_PPPY <dbl>
#> 
#> $`Summary by settings using gtsummary`
#> <div id="wkkeulbfdp" style="padding-left:0px;padding-right:0px;padding-top:10px;padding-bottom:10px;overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>#wkkeulbfdp table {
#>   font-family: system-ui, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
#>   -webkit-font-smoothing: antialiased;
#>   -moz-osx-font-smoothing: grayscale;
#> }
#> 
#> #wkkeulbfdp thead, #wkkeulbfdp tbody, #wkkeulbfdp tfoot, #wkkeulbfdp tr, #wkkeulbfdp td, #wkkeulbfdp th {
#>   border-style: none;
#> }
#> 
#> #wkkeulbfdp p {
#>   margin: 0;
#>   padding: 0;
#> }
#> 
#> #wkkeulbfdp .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   line-height: normal;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #wkkeulbfdp .gt_caption {
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #wkkeulbfdp .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #wkkeulbfdp .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 3px;
#>   padding-bottom: 5px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #wkkeulbfdp .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #wkkeulbfdp .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #wkkeulbfdp .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #wkkeulbfdp .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #wkkeulbfdp .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #wkkeulbfdp .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #wkkeulbfdp .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #wkkeulbfdp .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #wkkeulbfdp .gt_spanner_row {
#>   border-bottom-style: hidden;
#> }
#> 
#> #wkkeulbfdp .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   text-align: left;
#> }
#> 
#> #wkkeulbfdp .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #wkkeulbfdp .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #wkkeulbfdp .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #wkkeulbfdp .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #wkkeulbfdp .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #wkkeulbfdp .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #wkkeulbfdp .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #wkkeulbfdp .gt_row_group_first th {
#>   border-top-width: 2px;
#> }
#> 
#> #wkkeulbfdp .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #wkkeulbfdp .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #wkkeulbfdp .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #wkkeulbfdp .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #wkkeulbfdp .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #wkkeulbfdp .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #wkkeulbfdp .gt_last_grand_summary_row_top {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: double;
#>   border-bottom-width: 6px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #wkkeulbfdp .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #wkkeulbfdp .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #wkkeulbfdp .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #wkkeulbfdp .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #wkkeulbfdp .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #wkkeulbfdp .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #wkkeulbfdp .gt_left {
#>   text-align: left;
#> }
#> 
#> #wkkeulbfdp .gt_center {
#>   text-align: center;
#> }
#> 
#> #wkkeulbfdp .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #wkkeulbfdp .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #wkkeulbfdp .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #wkkeulbfdp .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #wkkeulbfdp .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #wkkeulbfdp .gt_footnote_marks {
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#>   position: initial;
#> }
#> 
#> #wkkeulbfdp .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #wkkeulbfdp .gt_indent_1 {
#>   text-indent: 5px;
#> }
#> 
#> #wkkeulbfdp .gt_indent_2 {
#>   text-indent: 10px;
#> }
#> 
#> #wkkeulbfdp .gt_indent_3 {
#>   text-indent: 15px;
#> }
#> 
#> #wkkeulbfdp .gt_indent_4 {
#>   text-indent: 20px;
#> }
#> 
#> #wkkeulbfdp .gt_indent_5 {
#>   text-indent: 25px;
#> }
#> 
#> #wkkeulbfdp .katex-display {
#>   display: inline-flex !important;
#>   margin-bottom: 0.75em !important;
#> }
#> 
#> #wkkeulbfdp div.Reactable > div.rt-table > div.rt-thead > div.rt-tr.rt-tr-group-header > div.rt-th-group:after {
#>   height: 0px !important;
#> }
#> </style>
#>   <table class="gt_table" data-quarto-disable-processing="false" data-quarto-bootstrap="false">
#>   <!--/html_preserve--><caption class='gt_caption'><span class='gt_from_md'><strong>Summary Table</strong></span></caption><!--html_preserve-->
#>   <thead>
#>     <tr class="gt_col_headings gt_spanner_row">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="2" colspan="1" scope="col" id="label"><span class='gt_from_md'><strong>Characteristic</strong></span></th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="1" scope="col" id="level 1; stat_1_1">
#>         <div class="gt_column_spanner"><span class='gt_from_md'>OP</span></div>
#>       </th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="1" scope="col" id="level 1; stat_1_2">
#>         <div class="gt_column_spanner"><span class='gt_from_md'>ED</span></div>
#>       </th>
#>     </tr>
#>     <tr class="gt_col_headings">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="stat_1_1"><span class='gt_from_md'>control, N = 2</span><span class="gt_footnote_marks" style="white-space:nowrap;font-style:italic;font-weight:normal;line-height:0;"><sup>1</sup></span></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="stat_1_2"><span class='gt_from_md'>treatment, N = 2</span><span class="gt_footnote_marks" style="white-space:nowrap;font-style:italic;font-weight:normal;line-height:0;"><sup>1</sup></span></th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td headers="label" class="gt_row gt_left" style="font-weight: bold;">Visits</td>
#> <td headers="stat_1_1" class="gt_row gt_center"><br /></td>
#> <td headers="stat_1_2" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    1</td>
#> <td headers="stat_1_1" class="gt_row gt_center">2 (100%)</td>
#> <td headers="stat_1_2" class="gt_row gt_center">2 (100%)</td></tr>
#>     <tr><td headers="label" class="gt_row gt_left" style="font-weight: bold;">Cost</td>
#> <td headers="stat_1_1" class="gt_row gt_center"><br /></td>
#> <td headers="stat_1_2" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    175.25</td>
#> <td headers="stat_1_1" class="gt_row gt_center">1 (50%)</td>
#> <td headers="stat_1_2" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    586.6</td>
#> <td headers="stat_1_1" class="gt_row gt_center">1 (50%)</td>
#> <td headers="stat_1_2" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="label" class="gt_row gt_left" style="font-weight: bold;">Visit_PPPM</td>
#> <td headers="stat_1_1" class="gt_row gt_center"><br /></td>
#> <td headers="stat_1_2" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    0.370939024390244</td>
#> <td headers="stat_1_1" class="gt_row gt_center">1 (50%)</td>
#> <td headers="stat_1_2" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    2.17264285714286</td>
#> <td headers="stat_1_1" class="gt_row gt_center">1 (50%)</td>
#> <td headers="stat_1_2" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="label" class="gt_row gt_left" style="font-weight: bold;">Visit_PPPY</td>
#> <td headers="stat_1_1" class="gt_row gt_center"><br /></td>
#> <td headers="stat_1_2" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    4.45731707317073</td>
#> <td headers="stat_1_1" class="gt_row gt_center">1 (50%)</td>
#> <td headers="stat_1_2" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    26.1071428571429</td>
#> <td headers="stat_1_1" class="gt_row gt_center">1 (50%)</td>
#> <td headers="stat_1_2" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="label" class="gt_row gt_left" style="font-weight: bold;">Cost_PPPM</td>
#> <td headers="stat_1_1" class="gt_row gt_center"><br /></td>
#> <td headers="stat_1_2" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    65.0070640243902</td>
#> <td headers="stat_1_1" class="gt_row gt_center">1 (50%)</td>
#> <td headers="stat_1_2" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    1274.4723</td>
#> <td headers="stat_1_1" class="gt_row gt_center">1 (50%)</td>
#> <td headers="stat_1_2" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="label" class="gt_row gt_left" style="font-weight: bold;">Cost_PPPY</td>
#> <td headers="stat_1_1" class="gt_row gt_center"><br /></td>
#> <td headers="stat_1_2" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    781.144817073171</td>
#> <td headers="stat_1_1" class="gt_row gt_center">1 (50%)</td>
#> <td headers="stat_1_2" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    15314.45</td>
#> <td headers="stat_1_1" class="gt_row gt_center">1 (50%)</td>
#> <td headers="stat_1_2" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    1312.47</td>
#> <td headers="stat_1_1" class="gt_row gt_center"><br /></td>
#> <td headers="stat_1_2" class="gt_row gt_center">1 (50%)</td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    4885.25</td>
#> <td headers="stat_1_1" class="gt_row gt_center"><br /></td>
#> <td headers="stat_1_2" class="gt_row gt_center">1 (50%)</td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    0.95053125</td>
#> <td headers="stat_1_1" class="gt_row gt_center"><br /></td>
#> <td headers="stat_1_2" class="gt_row gt_center">1 (50%)</td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    6.0834</td>
#> <td headers="stat_1_1" class="gt_row gt_center"><br /></td>
#> <td headers="stat_1_2" class="gt_row gt_center">1 (50%)</td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    11.421875</td>
#> <td headers="stat_1_1" class="gt_row gt_center"><br /></td>
#> <td headers="stat_1_2" class="gt_row gt_center">1 (50%)</td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    73.1</td>
#> <td headers="stat_1_1" class="gt_row gt_center"><br /></td>
#> <td headers="stat_1_2" class="gt_row gt_center">1 (50%)</td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    1247.5437496875</td>
#> <td headers="stat_1_1" class="gt_row gt_center"><br /></td>
#> <td headers="stat_1_2" class="gt_row gt_center">1 (50%)</td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    29718.92985</td>
#> <td headers="stat_1_1" class="gt_row gt_center"><br /></td>
#> <td headers="stat_1_2" class="gt_row gt_center">1 (50%)</td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    14990.86828125</td>
#> <td headers="stat_1_1" class="gt_row gt_center"><br /></td>
#> <td headers="stat_1_2" class="gt_row gt_center">1 (50%)</td></tr>
#>     <tr><td headers="label" class="gt_row gt_left">    357111.775</td>
#> <td headers="stat_1_1" class="gt_row gt_center"><br /></td>
#> <td headers="stat_1_2" class="gt_row gt_center">1 (50%)</td></tr>
#>   </tbody>
#>   
#>   <tfoot class="gt_footnotes">
#>     <tr>
#>       <td class="gt_footnote" colspan="3"><span class="gt_footnote_marks" style="white-space:nowrap;font-style:italic;font-weight:normal;line-height:0;"><sup>1</sup></span> <span class='gt_from_md'>n (%)</span></td>
#>     </tr>
#>   </tfoot>
#> </table>
#> </div>
#> 
# }