Generate Detailed Descriptive Statistics

Generate Detailed Descriptive Statistics

Usage

summarize_descriptives(
  data,
  patient_id_col = "patient_id",
  setting_col = "care_setting",
  cohort_col = "cohort",
  encounter_id_col = "encounter_id",
  cost_col = "cost_usd",
  los_col = "length_of_stay",
  readmission_col = "readmission",
  time_window_col = "period"
)

Arguments

data

A dataframe with variables to summarize.

patient_id_col

A character specifying the name of patient identifier column

setting_col

A character specifying the name of HRCU setting column

cohort_col

A character specifying the name of cohort column

encounter_id_col

A character specifying the name of encounter/claim column

cost_col

A character specifying the name of cost column

los_col

A character specifying the name of length of stay column

readmission_col

A character specifying the name of readmission column

time_window_col

A character specifying the name of time window column

Value

A table object